config/requirements.txt
# Flask and extensions
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1

# Database drivers
pymysql==1.1.0
pymongo==4.6.1

# Environment and configuration
python-dotenv==1.0.0

# Utilities
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
PyYAML==6.0.1
email_validator

# Data processing
pandas==2.1.1
numpy==1.26.0

# Testing
pytest==7.4.2
pytest-flask==1.3.0

# Development tools
black==23.9.1
flake8==6.1.0

# Docker support
docker==6.1.3